Added PoC search, String -> Text
Added an ability to search, ignoring case, based on the title. Still need to
make it handle paging at the same time, but that's a function of the links I'm
generating for the pages.

I also converted everything to "Text" (as much as I could) and chucked in a
few pack/unpacks.